Don't need the gmo file
git-svn-id: http://svn.nickg.me.uk/work/lander/trunk@345 a97b1542-0b21-0410-a459-e47997c36f34

+++ src/Mechanics.hpp
@@ -25,8 +25,8 @@ public:
 
    static Velocity Project(float speed, float angle);
 
-   int GetXSpeed() const { return xspeed; }
-   int GetYSpeed() const { return yspeed; }
+   float GetXSpeed() const { return xspeed; }
+   float GetYSpeed() const { return yspeed; }
    
 private:
    float xspeed, yspeed;
